feat(esp): LilyGo T-Display S3 AMOLED Plus bt2usb target with animated eyes
Add a bt2usb board target for the LilyGo T-Display S3 AMOLED Plus
(ESP32-S3R8, 16MB flash, 8MB OPI PSRAM, RM67162 1.91" 240x536 AMOLED).
The board enumerates as a Joypad USB device and drives animated companion
eyes on the AMOLED.

- Board config + 16MB partition. ESPTOOLPY_BEFORE_NORESET for the stock
  (non-TinyUF2) bootloader so esptool doesn't fight the reset; enter
  download with BOOT+RST.
- RM67162 single-SPI driver over esp_lcd: DC pin, GPIO38 PMIC power-enable
  (panel stays dark without it), full rm67162_spi_cmd init sequence
  (0x53=0x20 brightness-control enable is required). Async tx_color is
  guarded by an on_color_trans_done semaphore to avoid DMA-race tearing.
- Animated eyes: the shared eyes_anim engine, rendered by a FreeRTOS task
  and blitted to the panel. New EYES_SCALE compile factor (=4 here) renders
  the eyes at high resolution instead of a nearest-neighbor upscale;
  display_pixel widened to int16_t. RP2040 128x64 path is unchanged
  (EYES_SCALE=1) and still builds (verified usb2usb_feather_rp2040).
- platform_reboot_bootloader forces ROM download on this board (no TinyUF2)
  and self-heals the sticky force-download bit on the next boot.

+++ src/core/services/display/display.h
@@ -48,8 +48,10 @@ void display_clear(void);
 // Update display (send framebuffer to OLED)
 void display_update(void);
 
-// Set pixel at x,y (0=off, 1=on)
-void display_pixel(uint8_t x, uint8_t y, bool on);
+// Set pixel at x,y (0=off, 1=on). Coords are int16_t so a higher-resolution
+// backend (e.g. the AMOLED eyes canvas) can address beyond 255; the 128x64
+// OLED backend simply bounds-checks against its own dimensions.
+void display_pixel(int16_t x, int16_t y, bool on);
 
 // Draw text at position (using built-in 6x8 font)
 void display_text(uint8_t x, uint8_t y, const char* text);
